System.RSS

Documentation

currentRSS :: IO CSize

Retrieve the current resident set size for the currently executing program.

peakRSS :: IO CSize

Retrieve the current peak resident set size for the currently executing program.