System.Mem.Manager

Description

This module supplies a variant of "Poor Richard's Memory Manager" by Hertz, Kane, Keudel, Bai, Ding, Gu and Bard, adapted to run in Haskell in user-space.

To incorporate this into your program invoke selfishManager from your main to spawn a background thread that will perform garbage collection whenever it encounters signs of system back-pressure, rather than just when a predefined system memory limit is reached.

It'd be possible to implement communal and leadership whiteboarding between programs as well, but this should serve as a sufficient proof of concept for now.

Memory Manager

selfishManager :: IO ThreadId

Run a background thread that checks for signs of memory pressure from the Host OS and kickstarts a garbage collection as needed.

Statistics and Utilities

currentResidentSetSize :: IO CSize

Retrieve the current resident set size for the currently executing program.

peakResidentSetSize :: IO CSize

Retrieve the current peak resident set size for the currently executing program.

hardPageFaults :: IO CSize

Return the total number of "hard page-faults" since the program started. These are page-faults which required us to go out to disk.

checkMemoryPressure :: IO Bool

Determine if memory pressure warrants further action.

cumulativeManagedCollections :: IO Int

Return the total number of managed collections that the GC manager has forced.